refactor: move vote related code to `vote-registration` files
